+++ src/crisptrutski/boot_cljs_test.clj
@@ -1,10 +1,11 @@
 (ns crisptrutski.boot-cljs-test
   (:require
-   [clojure.java.io :as io]
-   [clojure.string :as str]
-   [boot.core :as core :refer [deftask]]
-   [boot.util :refer [info dbug warn fail]]
-   [crisptrutski.boot-cljs-test.utils :as u])
+    [boot.core :as core :refer [deftask]]
+    [boot.util :refer [info dbug warn fail]]
+    [clojure.java.io :as io]
+    [clojure.string :as str]
+    [crisptrutski.boot-cljs-test.utils :as u]
+   [crisptrutski.boot-error.core :as err])
   (:import
    [java.io File]))
 
@@ -12,164 +13,195 @@
   {:adzerk/boot-cljs "1.7.170-3"
    :doo              "0.1.7-SNAPSHOT"})
 
-(def default-js-env   :phantom)
-(def default-suite-ns 'cljs-test.suite)
-(def default-output   "output.js")
-
-;; state
-
-(def failures? (atom false))
+(def default-js-env :phantom)
+(def default-ids    ["cljs_test/suite"])
 
 ;; core
 
+(defn no-op [& _])
+
+(defn- scope-as
+  "Modify dependency co-ords to have particular scope.
+   Assumes not currently scoped"
+  [scope deps]
+  (for [co-ords deps]
+    (conj co-ords :scope scope)))
+
 (defn ensure-deps! [keys]
-  (core/set-env! :dependencies #(into % (u/filter-deps keys deps))))
+  (when-let [deps (seq (u/filter-deps keys deps))]
+    (warn "Adding: %s to :dependencies" deps)
+    (core/merge-env! :dependencies (scope-as "test" deps))))
+
+(defn validate-cljs-opts! [js-env cljs-opts]
+  ((u/r doo.core/assert-compiler-opts) js-env (assoc cljs-opts
+                                     :output-to "placeholder"
+                                     :output-dir "placeholder"
+                                                :assert-path "placeholder")))
 
 (defn- gen-suite-ns
   "Generate source-code for default test suite."
-  [ns test-namespaces]
-  (let [ns-spec `(~'ns ~ns (:require [doo.runner :refer-macros [~'doo-tests ~'doo-all-tests]]
-                                     ~@(mapv vector test-namespaces)))
-        run-exp `(~'doo-tests ~@(map u/normalize-sym test-namespaces))]
-    (->> [ns-spec run-exp]
+  [ns namespaces]
+  (let [ns-spec `(~'ns ~ns (:require [doo.runner :refer-macros [~'doo-tests]] ~@(mapv vector namespaces)))
+        run-exp `(~'doo-tests ~@(map u/normalize-sym namespaces))]
+    (->> [ns-spec '(enable-console-print!) run-exp]
          (map #(with-out-str (clojure.pprint/pprint %)))
          (str/join "\n" ))))
 
 (defn add-suite-ns!
   "Add test suite bootstrap script to fileset."
-  [fileset tmp-main suite-ns test-namespaces]
-  (ensure-deps! [:adzerk/boot-cljs])
-  (let [out-main (u/ns->cljs-path suite-ns)
-        out-file (doto (io/file tmp-main out-main) io/make-parents)
-        out-path (u/relativize (.getPath tmp-main) (.getPath out-file))
-        cljs     (u/cljs-files fileset)]
-    (if (contains? (into #{} (map core/tmp-path) cljs) out-path)
-      (info "Using %s...\n" out-path)
-      (do (info "Writing %s...\n" out-path)
-          (spit out-file (gen-suite-ns suite-ns test-namespaces))))))
+  [fileset tmp-main id namespaces verbosity]
+  (let [relative   #(u/relativize (.getPath tmp-main) (.getPath %))
+        out-main   (str id ".cljs")
+        src-file   (doto (io/file tmp-main out-main) io/make-parents)
+        edn-file   (io/file tmp-main (str out-main ".edn"))
+        src-path   (relative src-file)
+        edn-path   (relative edn-file)
+        exists?    (into #{} (map core/tmp-path) (u/cljs-files fileset))
+        suite?     (or (exists? src-path) (exists? (str/replace src-path ".cljs" ".cljc")))
+        edn?       (exists? edn-path)
+        edn        (when edn? (read-string (slurp (core/tmp-file (core/tmp-get fileset edn-path)))))
+        ;; TODO: better would be to use transitive requires, once we have the dep graph
+        ;; subset namespaces to those directly required by .cljs.edn
+        namespaces (if edn (filter (set (:require edn)) namespaces) namespaces)
+        suite-ns   (u/file->ns out-main)
+        info       (if (pos? verbosity) info no-op)]
+    (if suite?
+      (info "Using %s...\n" src-path)
+      (do
+        (info "Writing %s...\n" src-path)
+        (spit src-file (gen-suite-ns suite-ns namespaces))))
+    (if edn?
+      ;; ensure that .cljs file is required by .cljs.edn, if it's being created?
+      (if suite?
+        (info "Using %s...\n" edn-path)
+        (do (info "Updating %s...\n" edn-path)
+            (spit edn-file (update edn :require (fn [xs] (when-not (some #{suite-ns} xs)
+                                                           (conj xs suite-ns)))))))
+      (do (info "Writing %s...\n" edn-path)
+          (spit edn-file {:require [suite-ns]})))
+    (if (and suite? edn?)
+      fileset
+      (core/commit! (core/add-source fileset tmp-main)))))
 
 (deftask prep-cljs-tests
   "Prepare fileset to compile main entry point for the test suite."
-  [o out-file   VAL str       "Output file for test script."
-   n namespaces NS ^:! #{str} "Namespaces whose tests will be run. All tests will be run if
+  [n namespaces NS ^:! #{str} "Namespaces whose tests will be run. All tests will be run if
                                ommitted.
                                Use symbols for literals.
                                Regexes are also supported.
                                Strings will be coerced to entire regexes."
-   s suite-ns   NS  sym       "Test entry point. If this is not provided, a namespace will be
-                               generated."]
-  (let [out-file (or out-file default-output)
-        out-id   (str/replace out-file #"\.js$" "")
-        suite-ns (or suite-ns default-suite-ns)
-        tmp-main (core/tmp-dir!)]
+   e exclusions NS ^:! #{str} "Namespaces or namesaces patterns to exclude."
+   v verbosity VAL  int       "Log level"
+   i id         VAL str       "TODO: WRITE ME"]
+  (let [tmp-main (core/tmp-dir!)
+        verbosity (or verbosity @boot.util/*verbosity*)]
     (core/with-pre-wrap fileset
-      (let [namespaces (u/refine-namespaces fileset namespaces)]
+      (let [namespaces (u/refine-namespaces fileset namespaces exclusions)]
         (core/empty-dir! tmp-main)
-        (info "Writing %s...\n" (str out-id ".cljs.edn"))
-        (spit (doto (io/file tmp-main (str out-id ".cljs.edn")) io/make-parents)
-              (pr-str {:require [suite-ns]}))
-        (add-suite-ns! fileset tmp-main suite-ns namespaces)
-        (-> fileset (core/add-source tmp-main) core/commit!)))))
+        (add-suite-ns! fileset tmp-main id namespaces verbosity)))))
+
+(defn- info? [verbosity & args]
+  (when (pos? verbosity) (apply info args)))
+
+(defn run-cljs-tests! [ids js-env cljs-opts v exit? doo-opts fileset]
+  (err/with-errors!
+    (info? v "Running cljs tests...\n")
+    ((u/r doo.core/print-envs) js-env)
+    (doseq [id ids]
+      (when (> (count ids) 1) (info? v "• %s\n" id))
+      (let [filename (str id ".js")
+            output-to (u/find-path fileset filename)
+            output-dir (str/replace output-to #"\.js\z" ".out")
+            cljs-opts (u/build-cljs-opts cljs-opts output-to output-dir)]
+        ((u/r doo.core/assert-compiler-opts) js-env cljs-opts)
+        (if-not output-to
+          (do (warn "Test script not found: %s\n" filename)
+              (err/track-error! {:exit 1 :out "" :err (format "Test script not found: %s" filename)})
+              (when exit? (System/exit 1)))
+          (let [dir (.getParentFile (File. ^String output-to))
+                {:keys [exit] :as result}
+                ((u/r doo.core/run-script) js-env cljs-opts
+                  (merge
+                    doo-opts
+                    {:exec-dir dir
+                     :verbose (>= verbosity 1)
+                     :debug (>= verbosity 2)}))]
+            (when (pos? exit)
+              (err/track-error! result)
+              (when exit? (System/exit exit)))))))
+    fileset))
 
 (deftask run-cljs-tests
   "Execute test reporter on compiled tests"
-  [o out-file   VAL str  "Output file for test script."
-   e js-env     VAL kw   "The environment to run tests within, eg. phantom"
-   c cljs-opts  VAL code "Compiler options for CLJS"
-   d doo-opts   VAL code "Options to pass to doo"
-   x exit?          bool "Exit immediately with reporter's exit code."]
-  (let [js-env   (or js-env default-js-env)
-        out-file (or out-file default-output)]
-    (ensure-deps! [:doo])
-    ;;((r doo.core/assert-compiler-opts) js-env {:output-to out-file})
+  [i ids        IDS  [str] ""
+   j js-env     VAL  kw     "Environment to execute within, eg. slimer, phantom, ..."
+   c cljs-opts  OPTS edn    "Options to pass to the Clojurescript compiler."
+   v verbosity  VAL  int    "Log level"
+   d doo-opts   VAL  code    "Options for doo"
+   x exit?           bool   "Exit process with runner's exit code on completion."]
+  (ensure-deps! [:doo])
+  (let [js-env (or js-env default-js-env)
+        ids (if (seq ids) ids default-ids)
+        verbosity (or verbosity @boot.util/*verbosity*)]
+    (validate-cljs-opts! js-env cljs-opts)
     (fn [next-task]
       (fn [fileset]
-        (info "Running cljs tests...\n")
-        (if-let [path (some->> (core/output-files fileset)
-                               (filter (comp #{out-file} :path))
-                               (sort-by :time)
-                               (last)
-                               (core/tmp-file)
-                               (.getPath))]
-          (let [dir (.getParentFile (File. path))
-                ;; TODO: could infer :asset-path and :main too
-                ;; TODO: perhaps better to get boot-cljs to share this parsing logic
-                cljs (merge cljs-opts {:output-to path, :output-dir (str/replace path #".js\z" ".out")})
-                opts (merge doo-opts {:exec-dir dir})
-                {:keys [exit] :as result}
-                ((u/r doo.core/run-script) js-env cljs opts)]
-            (when (pos? exit) (reset! failures? true))
-            (when exit? (System/exit exit))
-            (next-task fileset))
-          (do (warn "Test script not found: %s\n" out-file)
-              (if exit? (System/exit 1) (next-task fileset))))))))
+        (next-task (run-cljs-tests! ids js-env cljs-opts verbosity exit? doo-opts fileset))))))
 
-(defn- capture-fileset [fs-atom]
-  (fn [next-task]
-    (fn [fileset]
-      (reset! fs-atom fileset)
-      (next-task fileset))))
-
-(defn- return-fileset [fs-atom]
-  (fn [next-task]
-    (fn [_]
-      (let [fileset @fs-atom]
-        (core/commit! fileset)
-        (next-task fileset)))))
+(defn -test-cljs
+  [js-env namespaces exclusions optimizations ids out-file cljs-opts verbosity update-fs? exit?]
+  (ensure-deps! [:adzerk/boot-cljs])
+  (when out-file (warn "[boot-cljs] :out-file is deprecated, please use :ids\n"))
+  (let [verbosity (or verbosity @boot.util/*verbosity*)
+        ids (if (seq ids) ids (if out-file (str/replace out-file #"\.js\z" "") default-ids))
+        optimizations (or optimizations :none)
+        js-env (or js-env default-js-env)
+        cljs-opts (u/combine-cljs-opts cljs-opts optimizations js-env)
+        wrapper (if update-fs? identity u/wrap-fs-rolback)]
+    (validate-cljs-opts! js-env cljs-opts)
+    (wrapper
+      (comp (reduce
+              comp
+              (for [id ids]
+                (prep-cljs-tests
+                  :id id
+                  :namespaces namespaces
+                  :exclusions exclusions)))
+            ((u/r adzerk.boot-cljs/cljs)
+              :ids (set ids)
+              :compiler-options cljs-opts)
+            (run-cljs-tests
+              :ids (vec (distinct ids))
+              :cljs-opts cljs-opts
+              :js-env js-env
+              :exit? exit?
+              :verbosity verbosity)))))
 
 (deftask test-cljs
   "Run cljs.test tests via the engine of your choice.
 
    The --namespaces option specifies the namespaces to test. The default is to
    run tests in all namespaces found in the project."
-  [e js-env        VAL   kw      "The environment to run tests within, eg. phantom"
-   n namespaces    NS ^:! #{str} "Namespaces whose tests will be run. All tests will be run if
-                                  ommitted."
-   s suite-ns      NS    sym     "Test entry point. If this is not provided, a namespace will be
-                                  generated."
-   O optimizations LEVEL kw      "The optimization level."
-   o out-file      VAL   str     "Output file for test script."
-   c cljs-opts     VAL   code    "Compiler options for CLJS"
-   d doo-opts      VAL   code    "Options for doo"
-   u update-fs?          bool    "Only if this is set does the next task's filset include
-                                  and generated or compiled cljs from the tests."
-   x exit?               bool    "Exit immediately with reporter's exit code."]
-  (ensure-deps! [:doo :adzerk/boot-cljs])
-  (let [out-file      (or out-file default-output)
-        out-id        (str/replace out-file #"\.js$" "")
-        optimizations (or optimizations :none)
-        js-env        (or js-env default-js-env)
-        suite-ns      (or suite-ns default-suite-ns)
-        cljs-opts     (merge {:main suite-ns, :optimizations optimizations}
-                             (when (= :node js-env) {:target :nodejs, :hashbang false})
-                             cljs-opts)
-        capture-atom  (atom nil)
-        ->fs          (if update-fs? identity (capture-fileset capture-atom))
-        fs->          (if update-fs? identity (return-fileset capture-atom))]
-    (if (and (= :none optimizations)
-             (= :rhino js-env))
-      (do
-        (fail "Combination of :rhino and :none is not currently supported.\n")
-        (if exit?
-          (System/exit 1)
-          identity))
-      (comp ->fs
-            (prep-cljs-tests :out-file out-file
-                             :namespaces namespaces
-                             :suite-ns suite-ns)
-            ((u/r adzerk.boot-cljs/cljs)
-             :ids              #{out-id}
-             :compiler-options cljs-opts)
-            (run-cljs-tests :out-file out-file
-                            :cljs-opts cljs-opts
-                            :doo-opts doo-opts
-                            :js-env js-env
-                            :exit? exit?)
-            fs->))))
+  [j js-env        VAL   kw      "Environment to execute within, eg. slimer, phantom, ..."
+   n namespaces    NS ^:! #{str} "Namepsaces or namespace patterns to run.
+                                  If omitted uses all namespaces ending in \"-test\"."
+   e exclusions    NS ^:! #{str} "Namespaces or namesaces patterns to exclude."
+   O optimizations LEVEL kw      "The optimization level, defaults to :none."
+   i ids           IDS   [str]  ""
+   o out-file      VAL   str     "DEPRECATED Output file for test script."
+   c cljs-opts     OPTS  edn     "Options to pass to the Clojurescript compiler."
+   u update-fs?          bool    "Enable fileset changes to be passed on to next task.
+                                  By default hides all effects for suite isolation."
+   v verbosity     VAL   int     "Verbosity level"
+   x exit?               bool    "Exit process with runner's exit code on completion."]
+  (ensure-deps! [:adzerk/boot-cljs])
+    (-test-cljs
+      js-env namespaces exclusions optimizations ids out-file cljs-opts verbosity update-fs? exit?))
 
 (deftask exit!
   "Exit with the appropriate error code"
   []
   (fn [_]
-    (fn [_]
-      (System/exit (if @failures? 1 0)))))
+    (fn [fs]
+      (let [errs (err/get-errors fs)]
+        (System/exit (if (seq errs) 1 0))))))